git-svn-id: file://localhost/home/tolik/svnrepo/tags/latest_tested_snapshot@1977 c5418bc3-7546-4aae-8bad-fa6f2f3e53e6

+++ opencv/src/cv/cvoptflowgf.cpp
@@ -42,7 +42,11 @@
 
 #include "_cv.h"
 
-// Gunnar Farneback 2D dense optical flow algorithm
+//
+// 2D dense optical flow algorithm from the following paper:
+// Gunnar Farneback. "Two-Frame Motion Estimation Based on Polynomial Expansion".
+// Proceedings of the 13th Scandinavian Conference on Image Analysis, Gothenburg, Sweden
+//
 
 namespace cv
 {
@@ -197,7 +201,7 @@ FarnebackUpdateMatrices( const Mat& _R0, const Mat& _R1, const Mat& _flow, Mat& 
 
     int x, y, width = _flow.cols, height = _flow.rows;
     const float* R1 = (float*)_R1.data;
-    int step1 = _R1.step/sizeof(R1[0]);
+    size_t step1 = _R1.step/sizeof(R1[0]);
     
     _M.create(height, width, CV_32FC(5));
